backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1
requests==2.32.3

